Fixes search's filter function (#29)
The built-in filter function used something built for searching for
files. This allowed for fuzzy matches, which was why I switched to it,
but it had weighted scoring that made no sense for this application. The
custom filter function now prefers exact matches and then falls back to
calculating Levenshtein distance. If the Levenshtein distance is too
great after entering enough search text, the template is discarded.

Closes #24